HumanEmotionsStreamlit/requirements.txt
streamlit
Pillow
torch
torchvision